Renaming `boundary_types.hpp` as `edge_types.hpp`
- cleaned up `#include` in `general_definitions.hpp`cc
